kybd: buffer scancodes intead of blocking
Previously, when we got a keyboard interrupt, we'd block until we read all the
scancodes needed for that interrupt (or timed out).  However, an interrupt is
actually generated for each scancode.

So now, we just buffer up the scancodes as we get them until we have a
'finished' buffer, at which point we cook it and generate a keypress event.