kthread: add kthread_pool thread pools
The thread pools consist of (a) a set of threads, and (b) a queue of
callbacks.  Callbacks (and associated args) are pushed onto the queue
from thread or interrupt contexts, waking up any idling worker threads.
The worker threads then process the queue.

This will probably not be used with n > 1 very often (since the kernel
is non-preemptive and single threaded, there isn't much point in a
multi-thread pool), but it's still a useful abstraction for a thread-
and interrupt-safe work queue.