isr: add fake stack frame to improve GDB backtraces
This creates a fake stack frame (consisting of the return address and
EBP) and pushes it onto the stack in the interrupt handler, after saving
all the registers.  It pops it on return.  GDB sees this stack frame,
which lets it find (e.g. in backtraces) the stack frame where the
interrupt actually occurred.