page_fault: fix asm output constraint for mov cr2
Previously, the compiler could generate a move into either a register or
a memory location, since the output was "=g".  GCC chose a register, so
it compiled file, but clang chooses memory, which is incorrect (CR2
can only be moved into a register).

This fixes the output constraint to only allow registers.