clang: add builtins.c with memcpy
Clang (and GCC) require certain function definitions to be available
when compiling with -ffreestanding.  memcpy is one such function.  GCC
doesn't (currently) generate calls to memcpy in our code, so it hasn't
been an issue.  Clang, however, does, which results in a link error.

We will likely have to add more functions to this file down the road, if
calls to other builtins are generated.