build: generate .d and .o files in build-out/
All intermediate build products (.o and .d files) are now generated in
the build-out/ directory, which is created as needed, to avoid
cluttering up the source tree with build by-products.

For the linker to find the manually-placed files referenced in the
linker script, we have to pass -Lbuild-out to ld.  For some reason, this
causes duplicate symbols if those files are also passed as normal
arguments (though if the object files are generated alongside the
source, and -L. is passed, there's no problem).  I'm not sure why this
is the case, but simply removing the manually-placed object files from
the argument list to ld fixes it and is harmless.