ext2: don't free inodes until all refs are closed
Previously, inodes would be greedily freed (in unlink() and rmdir()) if
the inode had no remaining links --- but some processes may still have a
file descriptor open on the file/directory.

Inodes are now not freed until the vnode_t is put, which requires a VFS
change to hold a reference to the child until after the unlink/rmdir is
finished.