block_cache: fix simultaneous flush bugs
This fixes several bugs.  First of all, multiple simultaneous flushes
were previously possible, which didn't work right.  Now, if an entry is
flushing, it cannot be put on the flush queue --- instead, threads
signal that they want a flush by resetting the flushed flag, and there
is a loop to keep flushing until the flag is no longer set.

Secondly, several places (specifically in maybe_free_cache_space) didn't
properly handle the currently-flushing case, and would free entries that
were in the process of being flushed.  This fixes those checks, and adds
a sanity check to make sure that when the write() in flush_cache_entry()
returns, the entry looks OK (and wasn't freed).