block_cache: ref memobj's for each of its entries
Previously, the memobj's refcount could drop to zero while there were
entries referring to it in the block cache.  This could cause
use-after-frees (for example, if the memobj was backed by a vnode, the
vnode might be freed, causing crashes if the entries were later
flushed).

Now, the memobj (and underlying resources, such as a vnode) will be kept
alive as long as there are cache entries in the block cache referencing
it.

Down the road, it may be desirable to add a mechanism for a memobj to
"force" the block cache to get rid of it's entries so that the
underlying resources can be freed.  This would be necessary for, e.g.,
unmounting a filesystem.