block_cache: do multiple passes in clear_unpinned()
This was causing a KASSERT failure when running the mmap tests followed
by the block_cache tests on ext2.  The reason is, in the second loop
(clearing the LRU queue), we would put() a memobj backed by a vnode in
the filesystem, which would in turn cause a get()/put() of block cache
entries corresponding to the filesystem metadata.  These would show up
in later iterations of the loop and fail KASSERT_DBG(entry->flushing).

Now, we make multiple passes through the flush queue and LRU queue,
clearing them in turn, until we reach a stable state of no flushable
entries.  This shouldn't loop forever as long as there are no block
cache entry "cycles".