block_cache test: set blocks to 0x0 in size test
This fixes a crash that occurs when the block cache test is run after
loading/exec'ing a process that calls fork().  I'm not sure why, but
that causes some block cache entries to stick around and keep the cache
size test from allocating new cache entries.  When it tries to put()
them, it crashes (since the pointers weren't initialized to 0x0).