syscall dmz: properly handle copying NULL buffers
For buffer arguments that can be NULL, we need to not allocate any
kernel memory for them, and not try to copy them into/out of the kernel
memory.  Likewise, we need to check the result of the kmalloc call iff
we actually tried to allocate kernel memory.

This changes allocation and checking for buffer arguments that can be
NULL to handle that case.  For *all* arguments, it now only does a
kmemcpy into/out of the buffer if the argument is non-NULL; this is a
no-op for required buffer/string arguments because if they are NULL,
they will be rejected earlier in the DMZ code.