interrupt test: zero upper 16 bits of pushed %cs
Bochs appears to only push the lower 16 bits of the %cs register,
leaving the upper 16 bits of the stack element unchanged.  This now
triggers a panic in the interrupt-handling code, which doesn't
recognize the fake CS pushed by the test.

This doesn't appear to be a problem with qemu or real hardware.