usb: add initial default address acquire/release
This will be used by the hub driver to ensure only one device is being
initialized with the default address at once.  Eventually, we'll need to
support queuing callbacks so that multiple hubs can interoperate.  For
the time being, this simply works as before (KASSERT'ing that the
default address is free) behind the new API.