memory: fix slab_alloc full-page accounting
The previous version didn't work correctly if you had an almost-full
page, then allocated the last object which wasn't in the last byte of
the bitmap.  We would need to look through the entire bitmap to
determine if the page was full.

This uses 3 additional bytes per page, but doesn't require a pass
through the bitmap.