build: add scons build files
The build system with scons is much easier to maintain and reason about
than the Makefiles, and will make it much easier to add other targets
(header distributions, userspace libraries, etc) down the road.  It is,
however, significantly slower: scons takes 1.5-2 seconds to realize that
nothing needs to be changed, while make can do it in <100ms.