proc: add interruptable waits for threads
A thread can now wait on a queue in an interruptable state, and another
thread can interrupt that waiting.  If that happens, the target thread
is removed from the queue it is waiting on and placed on the run queue,
setting a flag to indicate to the thread that it was interrupted.

The intended use of this is signals interrupting syscalls.