arch: fix fake stack frame for useful GDB backtraces
Switching from 'jmp' to 'call' to get to int_common_handler adds the
interrupt handler itself (e.g. int33) to the stack trace, which is
useful.  That makes back traces work in GDB for interrupts where the
error code is non-zero.

For interrupts where the error code is zero (either pushed by the CPU,
or the fake error code), GDB seems to get confused, think thats the
return address, and stop generating the stack.  Pushing the
interrupted EIP on *before* the interrupt number un-confuses GDB.

There are still one or two bogus stack frames in between the interrupted
code and the interrupt handler, but they're harmless.