syscall: make stack frame in _syscall_enter for traces
Previously, when the stack unwinder hit descended from syscall_dispatch
to _syscall_enter, it would pop whatever the value of %ebp was when
_syscall_enter was called, which could be anything.  It could either
segfault then, when trying to get the return address, or after reading a
bogus return address from the address in the popped %ebp.

Now, the popped %ebp will point to the beginning of the _syscall_enter
stack frame, so the return address will be the user-mode return address,
which the stack unwinder will recognize and stop at.