vterm: don't track line lengths
This is useful for backspace across newlines when the previous line
isn't the full with of the display.  In practice, that can't happen when
an ld is attached to the terminal, since backspace over a newline is
only possible if the line wraps automatically, which can only happen if
the previous line reaches the edge of the screen.

Removing it simplifies other aspects of the vterm code (like cursor
handling).