vterm: don't scroll at end of line
Previously, the terminal would scroll as soon as the cursor moved past
the last column on the line, which in effect made it impossible to set
the bottom right character of the terminal.

Now, the cursor can be in the 'right margin', meaning one past the end
of the current line.  The next character will be put at the start of the
next line (as before), but the scroll won't happen until that character
is typed.  Additionally, if that next character is a newline, it is
ignored (and only the wrap happens), preventing a double newline.