x86: fix interrupt user-mode context usage
Previously, this would grab the saved copy of $ebp above the saved
registers, which is actually the $esp from the start of the interrupt,
which is in kernel memory.  This only manifests if user-mode is
interrupted, then the context is extracted, and then later restored
(i.e. because of a signal delivery), which is not currently exercised by
the tests.

Additionally, the wrong instruction was used to restore the context
(lret instead of iret), copied incorrectly when this code was moved to a
dedicated asm file.  This causes a GP fault (interrupt 13) when the
context was restored.