kthread: fix timeout-while-on-run-queue crash bug
If a thread is woken up normally, but the timeout fires before it
actually gets a chance to run (and cancel the timeout), this would
previously fail an assertion that didn't distinguish between the
SWAIT_DONE states (pre-wakeup and post-normal-wakeup).  Now we simply
ignore the timeout in that scenario.

Something about the timing of the tests when run on x64 tickles this.