poll: fix timeout underflow bug on x64
end_time and now are both 32-bit unsigned integers.  If timeout_ms is
negative, (end_time - now) will underflow, resulting in a large signed
number.  On x86, when this is converted to a signed long to be passed to
scheduler_wait_on_interruptable(), it becomes -1 again.  On x64, it can
fit into a signed (64-bit) long, so a large positive number is passed as
the timeout instead of -1.

Arguably apos_ms_t shouldn't be uint32_t to begin with, but that's
another issue.