x64: add syscall context and support
This is mostly functional for 32-bit user binaries.  The problem is that
any syscalls with pointers to user data (mmap, anything that uses a
struct) has issues due to mismatched pointer and long sizes.

The long term solution will involve translating between the 32 and 64
bit versions of structures on entry and exit from the kernel.  In the
meantime, syscalls "work", but the user may get back corrupted data.  If
you rewrite all longs in user/syscall structs to be [u]int32_t,
the syscalls work great.

Note: needed to move the GDT descriptors around since the call gate
needs two entries, but also needs to be at the same offset as for i586
code.  So we bump the TSS down and leave slot 5 unused.