syscall: support 32-to-64 bit arg conversions
When running a 32-bit userspace binary with a 64-bit kernel, the sizes
of several fundemental types used in syscalls (such as long) will be
different.  If the value is passed directly, this isn't a problem, it
will be converted.

If the value is pointed to, however (e.g. as part of a stat struct, or a
signal handler function pointer), the kernel must be aware of the size
difference so it doesn't try to read/write 64-bit values from userpace
memory.

This adds a needs_32bit_conv parameter to syscall definitions that flags
such syscalls.  Each such syscall must define,
 (a) an explicit '32-bit' version of its arg types that's the same size
     on both 32 and 64-bit kernels.
 (b) a foo_32 syscall wrapper that converts to/from the 32-bit version
     of the pointed-to arguments and the native version of the type,
     and invokes the underlying native kernel handler.

The 32-bit wrappers and conversion functions must be written separately
for each type and syscall, but they could in principle be generated
automatically.

For now, in 64-bit kernels, all such incoming syscalls are treated as
though they're the 32-bit versions.  Once 64-bit userspace binaries are
supported, syscalls from 64-bit processes will be flagged as such, and
use the 'native' versions of the handlers (as 32->32 syscalls, and
non-size-dependent 32->64 syscalls do).