net: make new socks in connect instead of accept
Also fix the cleanup code.  This makes it much easier to deal with
closing sockets that have been connected but not accepted, since we can
just set the peer of the peer socket to NULL (rather than needing to
somehow remove it from the list, or have another object allocated that
can outlive the socket).

On OS X, calling accept() for a connection that's already closed doesn't
give client address info, so I replicate that here (it's easier
anyways).