vfs: fix cleanup on FIFO open() and pipe failures
Since vfs_open_fifo() can block, we have to allocate the file index and
file descriptor _after_ it returns---which means that if either of those
operations fails, we have to close the FIFO (currently we don't, which
causes an assertion failure).

This moves file/fd allocation down below the special per-file-type
handling as well as fixing cleanup. Nothing in the CHARDEV/TTY setup
blocks today, but this arrangement is clearer and prevents isuses if we
add blocking code there down the line.

This blocks fixing an rlimit issue, which blocks some more socket tests.