vfs: fix max-FDs rlimit overflow
By casting the limit (unsigned long) to int, we would overflow the limit
value if it was higher than INT_MAX.  This happened if the limit was
high, or all the time in 64-bit mode if the limit got clamped to
UINT32_MAX (from UINT64_MAX) in a getrlimit/setrlimit round-trip from
32-bit userspace.  This would result in max_fd being less than zero, and
we'd be unable to allocate a file descriptor.

Now we cast max_fd (the smaller type, int) to rlim_t before comparing to
the current rlimit, preventing overflows.