net: fix accept() buffer checking from userspace
Since the size of the buffer comes in another buffer argument, we need
to verify and dereference the addr_len argument before we can check the
addr argument.  This is further complicated by the fact that addr or
addr_len can be NULL.

I tried supporting this in the syscall templates, but it gets way too
complicated.  Write a custom wrapper instead.  Will need to do the same
for recvfrom().