x86/64: set invalid LDT at boot
Currently we get the default boot LDT which is at 0x0.  If user code
attempts an operation that references an LDT segment, we generate a
page-fault exception (in kernel mode) for the LDT entry in question.
This triggers an assertion failure at page_fault.c:57 (and would be a
bug regardless).

This explicitly loads an invalid LDT when we install the GDT.  This
causes the CPU (e.g. in a call instruction) to generate a GP fault if
the LDT is referenced, rather than trying to read it.