net: add net_link_send() and eth_send()
eth_send() is the interesting one---it takes a packet to be sent on an
ethernet NIC, does an ARP lookup, then sends it.  TBD if blocking is the
right behavior, but I'll keep it until I have a reason not to.

net_link_send() is just glue code to support other types of NICs (in
particular, loopback NICs).