net: set address in raw socket recvfrom()
Originally I considered extracting the address from the packet in the
raw socket code, which has the advantage of keeping the API smaller and
not needing the caller to extract the address (which gets hairy if e.g.
the IP code needs to extract an IP address vs a TCP address, etc).
However it requires the raw socket code to understand the different
protocols, which feels like a layering violation.

Assuming I go with the FreeBSD model where TCP, UDP, etc packets do
_not_ get returned on raw sockets then this works nicely (the IP code
can just fall back on the raw socket dispatch if it can't send the
packet to a more specialized protocol handler).