net/udp: implement poll() for UDP sockets.
It's unfortunate that pollable objects need to explicitly invalidate
their poll event by sending POLLNVAL; it would be nice if we had the
invariant (as for other references to the objects, e.g. in the rx wait
queue) that when cleanup() is called no other threads can be referencing
it.

Other socket types don't do this currently, I'll fix that in follow-up
commits.