proc: make deferred interrupt state per-thread
This makes the semantics of deferred interrupts match hardware
interrupts, and is required for code that needs to atomically read a
data structure then wait until it's updated by a defint (e.g. the ARP
cache).

The semantics aren't compatible with SMP, but any code using this will
have to be updated for SMP anyway.

This also simplifies the defint API a bit and properly prevents nested
defints from running.