proc: add basic non-SMP "spinlock"
This is obviously not actually a spinlock, since there's no SMP/SMT
enabled for the kernel yet, but models the behavior in a non-SMP system.
This gives a way to express a non-recursive preemption-disabling lock in
a way that will allow conversion to a preemptive kernel and be
future-proof for SMP support.