proc: disable defints in spinlocks
By disabling defints while spinlocks are held, we essentially treat
defints the same as any other thread, except they can preempt the
current thread at any time (rather than at the end of the current time
slice).  From a concurrency perspective, that makes them equivalent to
(preemptive) threads.

There's a perf hit, in that in theory code that doesn't need to
syncronize with defints is now disabling defints during the critical
section unnecessarily.  If that becomes a problem, it would be
straightforward to add another spinlock variant to differentiate in the
future.