proc: make spinlock interrupt-safety a propety of the lock
This is strictly speaking an overconstraint, but I'm having trouble
imagining a scenario in practice in which you'd want a particular lock
to operate in both interrupt-blocking and non-interrupt-blocking modes,
unless the code has some reason to know statically that there are some
times an interrupt might happen and other times it wouldn't.

In that case, I'd argue the code probably doesn't need the perf win, or
can split into two different locks.  That cost is outweighed by making
it harder to use the spinlock incorrectly, since whether or not it needs
to block interrupts becomes an immutable property of each lock rather
than a particular critical section.