vterm: spinlock for thread-safety/preemption
This is not strictly sufficient, since the vterm _can_ be accessed from
an interrupt handler.  However, given how much time is spent in vterm
code, disabling interrupts during all of that would be prohibitively
expensive.

For now, just hope that interrupts doing logging is rare enough to not
cause any problems (this is an existing issue, this commit does not make
it any worse).  A more nuanced locking structure that keeps the critical
sections small could reduce the cost considerably.