vfs: return vfs_get_memobj() object with refcount
This is not a bug today, but would be in the future (with
preemptible-mmap or SMP).  The vfs_get_memobj() call in mmap is what
binds the fd to a particular underlying vnode.  In the current code,
that is racy, and the vnode (and its memobj) could be destroyed before
we reference it if another thread closes the fd.

By moving the ref operation into vfs_get_memobj(), we force the
atomicity onto VFS internals (which are still racy today, of course) and
the caller is guaranteed their memobj will always be valid until they
unref it.