memory: make memobjs thread-safe with spinlocks
The only mutable part of the memobjs is the refcount, so that's all that
needs to be protected.  It's unfortunately challenging to refactor the
refcounting code into a shared memobj_{ref,unref}() implementation due
to how memobj_vnode works --- it's embedded in the owning vnode, and
vfs_put() can block (meaning we can't hold a spinlock across it), which
means we have to translate each memobj_vnodef ref/unref into an
underlying vnode ref/unref or be racy.

The refcount for vnode (and block device) memobjs is perfunctory, but
I'm keeping it for now so that at least it won't be surprising e.g. when
debugging to see a weird value.