proc: scheduler_wait_on_locked() for condition variables
This new variant lets us atomically unlock a mutex and wait on a thread
queue, which is needed for preemption and SMP support in most code that
uses scheduler_wait_on*().  It's possible we'll want a spinlock variant
as well.

The implementation is not SMP-safe, but that's now a problem contained
to the scheduler/kthread code.