memory: fix bug in block cache flushing unpinned entry
This bug was introduced in 2013 in commit be17f72d.  If an entry is on
the LRU queue but is unflushed, then is gotten again and put with
BD_FLUSH_NONE, then the next flush cycle will cause a KASSERT failure
because the block will have been left on the flush queue but marked as
flushed (incorrectly).

There's no reason to mark currently-unused entries as flushed when
they're gotten again; either they are in fact already flushed (in which
case flushed will be true and they won't be on the flush queue), or
they're still pending a flush (which will happen soon).

Ironically the original commit added a TODO for _exactly_ the test that
would have caught this.  Oops.