block cache: make preemption-safe
The model is a single global mutex that protects all state.  This
requires some careful unlocking/locking dances around calls into the
memobj that may either block or call back into the block cache.  This
has revealed a few latent bugs (fixed in previous commits).

If necessary in the future the lock (and other global state) could be
trivially sharded for performance; the only code that actually needs to
examine all global state is the periodic flush thread, which can iterate
over all shards (that is, there is no true global dependencies between
block cache objects).

The new code structure is a bit funny due to the need to unlock before
calling into the memobj.  This maps directly onto the non-preemptive
semantics (where state must be kept consistent before and after calls to
functions that can block), but is fragile.