ps2: process keyboard scanbuf in a defint
The actual keyboard processing isn't that expensive, but this is better,
and (more importantly) means that the vterm code doesn't need to be
interrupt-safe --- currently, any keypresses while the klog is printing
will cause a panic (before the spinlocks in vterm code, it would have
just had undefined but probably harmless behavior).

Any logs or assertion failures from interrupt handlers will still cause
a panic in the vterm code (perhaps overzealously), so there's still work
to be done.