test: clean up grandchildren in wait_tests
On i586-gcc (under qemu), this cleanup (in particular freeing the large
vm areas in the children) takes a very long time (~1.5s!).  On clang
it's fast (<50ms).

This was the case previously, but that cost was silently absorbed during
subsequent tests which would yield to thread 0 and block for 1.5s to
complete that cleanup.

After commit e807850, that cost falls to a poll test ("poll(): delayed
trigger wake up but no event test") and throws off the timing, causing
it to fail.

This commit cleans up the grandchilden synchronously in the test after
each use.  It is still slow on the GCC version, but nowthe cost is paid
by the test causing the problem (wait_test) and it doesn't inflict long
waits on other tests or cause them to spuriously fail.