vfs: add spinlock for vnode cache
The new test finds preemption races when preemption is enabled at the
given point.  This is not general thread-safety for VFS, only for the
vnode cache data structure.  In particular, other global data structures
(e.g. the file table) and per-vnode data structures (e.g. the state) are
not properly protected.