build: replace {'%s' % env['X']} with {'$X'}
This defers expansion of the variable until the end, once all build
files have been consumed.  That means that a sub-SConscript (e.g. in a
particular arch directory) can override construction environment
variables and have it apply at the top level, because the strings aren't
expanded until the end.

This is a no-op with all current build files (generated command lines
are identical).