common: switch to kfoo variants of standard types
For the common POSIX types, I went with a three-way variant (apos_foo,
kfoo, and foo).  kfoo is too short to be comfortable defining for all
user code, so apos_foo is the shared version.  Kernel code uses kfoo for
brevity.

Leaving ssize_t alone for now because it should ideally be handled the
same as size_t.