user: move signal constants to _posix_signal_constants.h header
Other than APOS_SIG{NULL,MIN,MAX}, these are all standard symbols.
Putting them in a dedicated header will make it easier to not define
them when we don't want them (for example, in a kernel-as-user-library
mode).

Leaving the APOS_* constants in place for now as their definition is
closely related to the actual signal numbers.