build: generate compile_commands.json db
The main kernel env triggers the compilation database generation, but it
includes compilation from all envs.  For that to work, each env must
load the tool, so we do that in base_env in SConstruct.  If multiple
envs call CompilationDatabase(), then we get duplicate entries.  If we
call CompilationDatabase() in SConstruct (e.g. on base_env), then the
database is emitted at the root of the project tree rather than in
build-scons/$ARCH-$TOOL/.