ext2: mark superblock const-after-creation fields
Ideally these could just be const, but I'm not sure that doesn't hit
undefined behavior.  A future change might be to put these fields into a
dedicated second struct that can be passed around independently of the
ext2fs_t or superblock --- enforcing structurally that these fields
aren't modified, and that code that depends on reading them without a
lock held doesn't accidentally introduce a read on another field that is
mutable.

Some of these are set when the filesystem is initally created.  Others
are in theory mutable, but we don't currently change them.