ext2: add lock, and use in sb/bg flushing
This lock will protect the in-memory copies of the global data
structures (superblock, block group descriptors, and block group
block/inode tables).

The way we flush is a bit odd, but this minimizes the length of the
critical section --- just enough to atomically capture the current state
of the in-memory copy, fix endianness (if necessary), and copy to the
disk block.  The IO (get and put) are done outside the critical section.