ext2: fix i_blocks accounting when freeing blocks
The previous commit _reset_ i_blocks, which is incorrect (and doesn't
account for indirect blocks).  Now we track how many blocks were freed,
and update i_blocks based on that.

This could be done with a closed-form calculation rather than tallying
as we go.  A future improvement to block accounting overall.

To repro:
 1) create large file (at least double-indirect)
 2) truncate to smaller file, ensuring that an indirect block is
    released.