ext2: don't run ltoh() on on-disk inode struct
This is far less important than for the global data structures, because
there's no possibility of concurrent access (if VFS does its job).
However, doing it this way (a) prevents us from ever writing an
"invalid" inode to disk (e.g. if it got flushed and the system crashed
before endianness could be corrected), and it avoids doing the ltoh() on
little-endian systems.

I'm tempted just to have this #error on non-little-endian systems,
because I can't actually test this right now.