vfs: remove no-op resolve_mounts_up in vfs_open()
This is extraneous with the call in lookup(), which is in turn called in
lookup_path_internal() on the final element of the path.

This was introduced in commit 7f81a8c6ffd429ff58c09b2a17ccc5ce3a0af7e3,
where the short-circuit return pre-lookup() for the final element was
removed.

Added a test that would catch this _not_ being present if it were also
removed from lookup_path() (otherwise only the lstat() path is tested
for mount resolution).