vfs: fix file deref race in vfs_close()
file_unref() is allowed to block --- which means another thread could
come along and observe the partially-closed file descriptor.

Longer-term the file table needs to be locked, which would also force
solving this.  Found this by unconditionally forcing a scheduler_yield()
in kmutex_lock() (to flush out this sort of issue).