vfs: use lookup_path()'s child in vfs_open()
The original child-lookup in vfs_open() is very old.  In commit
7a30a8defb8055e1fb88552d36deb71c576cd0fa, the lookup_path() call (which
didn't look up a child) was replaced with lookup_path2(), which had the
ability --- but vfs_open() wasn't updated to leverage that, because
lookup and creation needs to be atomic when the file is non-existent for
O_CREAT (without O_EXCL) to work properly.

This change resolves a race where the second lookup of the child could
result in getting a different child than the original lookup --- which
is probably fine, in general, except if the file is a symlink (which the
new test exercises).  Now, vfs_open() leverages the symlink and mount
resolution that happens in lookup_path_internal() and doesn't need to be
duplicated here.

lock_on_noent unfortunately isn't easily generalizable, and nothing
besides vfs_open() needs it (because it has hybrid open/create
semantics).