vfs: add lookup_path_and_lock() helper
This will be used for multi-node operations like unlink() and rmdir().

This version looks up the path and returns the parent and child both
locked.  To do this it has to do the initial lookup, lock the nodes with
vfs_lock_vnodes(), then confirm that the original lookup is still valid
(i.e. that it yields the child).

This two-step process (vs a more straightforward approach where we lock
the parent and then the child) is necessary to ensure there's no
deadlock with other threads that might lock the nodes in the opposite
order (e.g. as part of directory traversal, or a rename operation).

This does not ensure atomic of the full path, but that isn't necessary.
It also suffers from the ABA problem, but that's harmless (so long as
the right permissions still exist for the second lookup --- TODO left).

This could be easily adapted to support stabilizing on a lack-of-child
(e.g.  returning the parent locked and *child_out set to NULL, rather
than returning -ENOENT).  That said, none of the anticipated uses today
require that, so leaving it out for now.