proc: create kernel threads unattached to any process
This fixes an interesting and subtle bug: kernel threads that are
created late (for example, the block cache flush queue thread) will be
attached to whatever process is running at the time, which may later
exit.

This is triggered by commit 817ca506bd49e1880e3745414ef09d88910cb578, if
the root fs is set to be a ramfs (no attached disk) and the mount test
is run ASAP after boot.  In the new mmap_same_vnode_test(), when the
mmap is accessed, this triggers the first call into the block cache,
which in turns triggers initialization (and spawning the flush queue
thread).

Later, after the test exits, when the flush queue runs again, its
reference to its process points to garbage --- and this triggers
KASSERT-failure when the idle thread tries to switch to it.

A better fix might be to require callers to explicitly select a process
to own the new thread.  I don't want to update almost 100 call sites in
tests, though.  There are only four thread creation calls in non-test
code, three of which (block cache, kthread pool, scheduler) truly want
an unattached kernel thread (or one attached to the root process).  The
fourth, fork(), explicitly overrides the new thread's process after
creation.

It's still suboptimal that the block cache can be initialized from an
interrupt context.  I _think_ this is safe in practice, but may be
better to move to explicit initialization anyway.