vfs: set creation metadata (uid/gid) on symlinks
These aren't used, per se, but do seem to be populated on other systems
and are also returned by stat --- so we shouldn't leave them empty or as
garbage.

This was disguised by running tests on ext2, since the inode data is
cleared to zero in ext2, and there were no tests exercising non-zero
(superuser) uid/gid for symlinks.  Some existing tests failed on
ramfs, which flagged this bug.