vfs: make vfs_unlink() non-racy
It is now guaranteed that both the parent and child are locked when
fs->unlink() is called.  This guarantees that the unlinked entry is the
one that was originally looked up.

Writing the test to exercise this was very fiddly.  It required making
both ramfs and ext2 flag missing entries or other inconsistencies.