vfs: make vfs_rmdir() non-racy
As with vfs_unlink(), we now lock both the parent and child before
attempting to remove the directory.  The guarantees that no other thread
can either unlink/rmdir the child simultaneously, or add entries to it
(making it non-empty).

Another thread _may_ be traversing the path, or have its cwd inside the
child.  This is allowed (and I should write a test for it).