ramfs: track last allocated inode to cycle them
This is not a functional difference, but makes it easier to see what's
going on in a VFS rename test I'm working on.  Arguably this is worse
for testing, because we aren't as likely to exercise the inode-reuse
paths, but I think that's fine --- that logic should be sufficiently
covered by unit tests and not need to be exercised in all the main
functions.