vfs: make vfs_rename() non-racy.
The big kahuna!

The approach here is similar to the one for other multi-node operations
(unlink and rmdir), except that it requires locking up to four vnodes
(parent1, child1, parent2, child2) during lookup stabilization.

The filesystem-wide rename lock ensures that the overall topology
doesn't change while we stabilize lookups and execute the rename.