vfs: make vfs_link() non-racy
link() is mostly like other normal creation operations (mkdir, mknod,
etc), except that it's technically a 3-node operation (original parent,
new parent, node-to-be-linked).  This means it can't leverage
the standard locking behavior.

We don't lock the original parent because once the lookup happens, we
don't care about the binding to the child remaining valid.  If another
thread unbinds the child after the lookup, that's fine --- we hold an
(in-memory) reference, and so the link will succeed and be valid.