vfs: assert that multi-vnode locks are the same fs
Otherwise we risk deadlock --- imagine a filesystem that in turn is
backed by another filesystem (e.g. mounted from a block device that is
backed by a file on a different filesystem).  That means that fs
operations, including get_vnode(), might involve calls _back_ into VFS
code.  If we ensure that all multi-node locks other than the special
case of an uninitialized vnode go through these helpers, this helps
ensure we can never deadlock due to this situation.

In that case, the only cross-fs multi-vnode locks would be during a
vfs_get().  If the filesystem dependency graph is acyclic, and we only
ever hold locks on one filesystem at a time (other than vfs_get()),this
can't deadlock.  If there is a cycle in the filesystem dependency graph,
then this could deadlock, but that's just an impossible scenario to
handle correctly to begin with.