vfs: rewrite vfs_get()/put() and fix race
This serves several purposes,
1) fix a race when a thread calls vfs_get() for a vnode that another
   thread is in the process of vfs_put()ing
2) fix future data race on vnode->type, used to determine if the thread
   is initialized in vfs_get()
3) prepare for full thread-safety through VFS code.

In (1), the vnode is currently removed from the vnode table before
calling fs->put_vnode() on it.  If another thread calls vfs_get() for
that vnode while fs->put_vnode() is running, fs->get_vnode() is run
simultaneously, which breaks the contract with the fs impl.  Solving
this is tricky: we cannot handle it like block_cache and "flush" the
vnode asynchronously (and possibly reuse in a get/put race), because the
filesystem may reclaim the inode altogether during the put_vnode() call
if it is no longer referenced.  So the thread calling vfs_get() _must_
wait for the extant put_vnode() call to finish, then call
fs->get_vnode() to backfill from the fs again.

However, vfs_get() also can't take the vnode's mutex unless it knowns
the vnode is uninitialized, because that would introduce deadlocks if
vfs_get() is called while another vnode's lock is held.  The current
code examines vnode->type without a lock to accomplish this, but unless
type is made atomic, this is incompatible with making VFS
preemptible/thread-safe.

To address both of these, vfs_get() is split into two steps: binding and
initializing.  Binding binds the (fs, inode) tuple to a particular
identity (and vnode_t object), and doesn't require taking any locks
other than the vnode table spinlock.  Initializing actually fills the
vnode with data from the filesystem, which is a blocking operation.

The original plan was to expose these operations to other VFS code to
use, and code that needed to do a vfs_get() while holding another node's
lock (e.g. rename, lookup_existing_and_lock) would get an uninitialized
vnode, compare identity, then explicitly initialize if necessary
_without_ the other locks held.

This is unnecessary, however: if we maintain the property that vfs_get()
and vfs_put() only take the node's mutex (or block) when the vnode is
uninitialized/initializing and being destroyed, respectively, then we
cannot deadlock --- to deadlock, another thread would have to have the
vnode locked and be attempting to lock a second vnode.  In that case,
though, the vnode must be fully initialized (otherwise the other thread
couldn't have it) and we must not be destroying it (because the other
thread has a refcount on it).  I've kept the two operations separately
in the implementation for clarity.

This makes vfs_get() and vfs_put() more complicated, but localizes that
complexity --- other VFS code will only ever see initialized vnodes, and
can call vfs_get() and vfs_put() whenever without worrying about
deadlock.

To accomplish this, we track the vnode state explicitly and add another
lock which protects just that state, independently of the vnode's
data+metadata (which is protected by the current lock, and used by other
VFS code).  The refcount is protected by the global table lock --- there
is some potential future optimization to make this an atomic and only
take the table lock when necessary.

For the race test: to make this test effective, we need ramfs
jut_vnode() to take longer than get_vnode(), so we add another
maybe_block() call.  This is a bit clumsy, but means the bug reproduces
very reliably when running against ramfs (it does not against ext2).
Doing that in a more elegant way is left as a future exercise.